# openapi.yaml
...
openapi: 3.0.3
info:
  title: Oficina MVP API
  description: Tech Challenge 15SOAT — Sistema de Gestão de Oficina Mecânica (fases
    1 e 2).
  version: 1.0.0
...
paths:
...
  /ordens-servico:
    get:
      tags:
      - Ordens de Serviço
      summary: Lista Ordens de Serviço com filtros opcionais
      parameters:
      - name: clienteId
        in: query
        schema:
          $ref: "#/components/schemas/UUID"
      - name: status
        in: query
        schema:
          type: string
      - name: veiculoId
        in: query
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrdemServicoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    post:
      tags:
      - Ordens de Serviço
      summary: "Abertura de OS: recebe dados do cliente, veículo, serviços e peças\
        \ e retorna a identificação única da OS (cliente/veículo novos são cadastrados)"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbrirOrdemServicoRequest"
      responses:
        "201":
          description: OS criada (status RECEBIDA)
        "404":
          description: Serviço ou peça não encontrado
        "422":
          description: Estoque insuficiente em alguma peça
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}:
    get:
      tags:
      - Ordens de Serviço
      summary: "Detalhe completo da OS (itens, orçamento, datas)"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: OS não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/diagnostico:
    post:
      tags:
      - Ordens de Serviço
      summary: Inicia diagnóstico (RECEBIDA → EM_DIAGNOSTICO)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "409":
          description: Transição inválida
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/entregar:
    post:
      tags:
      - Ordens de Serviço
      summary: Atendente entrega o veículo (FINALIZADA → ENTREGUE)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "409":
          description: Transição inválida
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/finalizar:
    post:
      tags:
      - Ordens de Serviço
      summary: Mecânico finaliza serviços (EM_EXECUCAO → FINALIZADA)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "409":
          description: Transição inválida
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/orcamento:
    post:
      tags:
      - Ordens de Serviço
      summary: Gera orçamento e reserva peças no estoque
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "422":
          description: Estoque insuficiente em alguma peça
        "409":
          description: Transição inválida ou orçamento já gerado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/orcamento/aprovar:
    post:
      tags:
      - Ordens de Serviço
      summary: Aprova orçamento e dá baixa nas reservas (registro interno da decisão)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "409":
          description: Transição inválida ou orçamento inexistente
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/orcamento/enviar:
    post:
      tags:
      - Ordens de Serviço
      summary: Envia orçamento para aprovação (EM_DIAGNOSTICO → AGUARDANDO_APROVACAO)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "409":
          description: Transição inválida ou orçamento ainda não gerado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/orcamento/recusar:
    post:
      tags:
      - Ordens de Serviço
      summary: "Recusa orçamento: cancela a OS e libera as reservas de peças"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "409":
          description: Transição inválida ou orçamento inexistente
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/pecas:
    post:
      tags:
      - Ordens de Serviço
      summary: "Adiciona item de peça (valida saldo disponível, sem reservar)"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InserirItemPecaRequest"
      responses:
        "404":
          description: Peça ou OS não encontrada
        "422":
          description: Estoque insuficiente
        "409":
          description: Status incompatível ou orçamento já gerado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/pecas/{itemId}:
    delete:
      tags:
      - Ordens de Serviço
      summary: Remove item de peça (somente antes de gerar orçamento)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: itemId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServicoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/servicos:
    post:
      tags:
      - Ordens de Serviço
      summary: "Adiciona item de serviço (somente em EM_DIAGNOSTICO, sem orçamento)"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InserirItemServicoRequest"
      responses:
        "404":
          description: Serviço ou OS não encontrado
        "409":
          description: Status incompatível ou orçamento já gerado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/servicos/{itemId}:
    delete:
      tags:
      - Ordens de Serviço
      summary: Remove item de serviço (somente antes de gerar orçamento)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: itemId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServicoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /ordens-servico/{id}/status:
    get:
      tags:
      - Ordens de Serviço
      summary: "Consulta de status da OS: situação atual com descrição amigável"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: OS não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
...
  /servicos/{id}:
    get:
      tags:
      - Serviços
      summary: Busca um serviço pelo ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: Serviço não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    put:
      tags:
      - Serviços
      summary: Atualiza um serviço
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarServicoRequest"
      responses:
        "404":
          description: Serviço não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    delete:
      tags:
      - Serviços
      summary: Remove um serviço
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Removido
        "404":
          description: Serviço não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
...
components:
  schemas:
    AbrirOrdemServicoRequest:
      required:
      - cliente
      - veiculo
      type: object
      properties:
        cliente:
          $ref: "#/components/schemas/ClienteInput"
        veiculo:
          $ref: "#/components/schemas/VeiculoInput"
        servicos:
          type: array
          items:
            $ref: "#/components/schemas/ItemServicoInput"
        pecas:
          type: array
          items:
            $ref: "#/components/schemas/ItemPecaInput"
...
    AtualizarServicoRequest:
      required:
      - descricao
      - valorBase
      type: object
      properties:
        descricao:
          description: Descrição do serviço
          maxLength: 200
          pattern: \S
          type: string
          example: Troca de Óleo
        valorBase:
          description: Valor base do serviço
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 120.0
...
    ClienteInput:
      required:
      - documento
      type: object
      properties:
        documento:
          description: Documento (CPF/CNPJ) do cliente
          pattern: \S
          type: string
          example: "12345678900"
        nome:
          description: Nome — obrigatório se o cliente ainda não existir
          type: string
          example: Ana Souza
        email:
          description: E-mail para notificações de status da OS
          type: string
          example: ana@email.com
        telefone:
          type: string
          example: "11999990000"
...
    InserirItemPecaRequest:
      required:
      - pecaId
      - quantidade
      - valorUnitario
      type: object
      properties:
        pecaId:
          description: ID da peça
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
          example: 123e4567-e89b-12d3-a456-426614174000
        quantidade:
          format: int32
          description: Quantidade de peças
          minimum: 0
          exclusiveMinimum: true
          type: integer
          example: 2
        valorUnitario:
          description: Valor unitário cobrado pela peça
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 150.5
    InserirItemServicoRequest:
      required:
      - servicoId
      - valorCobrado
      type: object
      properties:
        servicoId:
          description: ID do serviço
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
          example: 123e4567-e89b-12d3-a456-426614174001
        valorCobrado:
          description: Valor cobrado pelo serviço
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 250.0
    ItemPecaInput:
      required:
      - pecaId
      - quantidade
      type: object
      properties:
        pecaId:
          description: ID da peça no estoque
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
          example: 123e4567-e89b-12d3-a456-426614174000
        quantidade:
          format: int32
          minimum: 0
          exclusiveMinimum: true
          type: integer
          example: 2
    ItemPecaView:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        pecaId:
          $ref: "#/components/schemas/UUID"
        quantidade:
          format: int32
          type: integer
        valorUnitario:
          type: number
        reservaId:
          $ref: "#/components/schemas/UUID"
    ItemServicoInput:
      required:
      - servicoId
      type: object
      properties:
        servicoId:
          description: ID do serviço no catálogo
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
          example: 123e4567-e89b-12d3-a456-426614174000
        valorCobrado:
          description: "Valor cobrado — se omitido, usa o valor base do catálogo"
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 350.0
    ItemServicoView:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        servicoId:
          $ref: "#/components/schemas/UUID"
        valorCobrado:
          type: number
...
    OffsetDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50-04:00
    OrcamentoView:
      type: object
      properties:
        valorTotal:
          type: number
        geradoEm:
          $ref: "#/components/schemas/OffsetDateTime"
        aprovadoEm:
          $ref: "#/components/schemas/OffsetDateTime"
    OrdemServicoResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        clienteId:
          $ref: "#/components/schemas/UUID"
        veiculoId:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
        itensServico:
          type: array
          items:
            $ref: "#/components/schemas/ItemServicoView"
        itensPeca:
          type: array
          items:
            $ref: "#/components/schemas/ItemPecaView"
        orcamento:
          $ref: "#/components/schemas/OrcamentoView"
        criadaEm:
          $ref: "#/components/schemas/OffsetDateTime"
        diagnosticoIniciadoEm:
          $ref: "#/components/schemas/OffsetDateTime"
        execucaoIniciadaEm:
          $ref: "#/components/schemas/OffsetDateTime"
        finalizadaEm:
          $ref: "#/components/schemas/OffsetDateTime"
        entregueEm:
          $ref: "#/components/schemas/OffsetDateTime"
        canceladaEm:
          $ref: "#/components/schemas/OffsetDateTime"
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
    VeiculoInput:
      required:
      - placa
      type: object
      properties:
        placa:
          description: Placa do veículo
          pattern: \S
          type: string
          example: ABC1234
        marca:
          description: Marca — obrigatória se o veículo ainda não existir
          type: string
          example: VW
        modelo:
          description: Modelo — obrigatório se o veículo ainda não existir
          type: string
          example: Golf
        ano:
          format: int32
          description: Ano — obrigatório se o veículo ainda não existir
          type: integer
          example: 2020
...
  securitySchemes:
    BearerAuth:
      type: http
      description: "JWT emitido pelo POST /auth/login (cole apenas o token, sem prefixo)."
      scheme: bearer
      bearerFormat: JWT
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT